Don't use fixture for testing post parser
If we write loads of assertions on data that's hidden away in an XML
file, we find ourselves writing hard to understand and unmaintainable
tests.

I'd rather write unit tests using data visible in the same file, using
hand crafted markup. A small number of integration tests can run on a
real export.