Quote page titles
Some of the posts in the site I'm converting contained characters (e.g.
colons) that broke the YAML parsing in Astro.

Wrapping the title in double quotes is an obvious solution, but that
broke on titles that contained a double quote character (they were
ending the string, as far as YAML was concerned).

So I'm quoting all titles, escaping double quotes within the titles.