Find attachment id of all inline images
We need to know which attachments to download for each post.

We've got a dictionary of attachments, mapping the id to the URL. By
finding the id of the attachments on each post, we'll be able to find
all the URLs.

We could just retrieve the URLs from the posts at this point (they're in
the <img> tags), but the site I'm working on uses gallery blocks [1].

Posts that have galleries on them are exported with just the attachment
ids in the content (without URLs), so we need to iterate over the ids
when downloading those images.

So collating ids for both inline and gallery images allows us to re-use
some logic.

[1] https://wordpress.com/support/wordpress-editor/blocks/gallery-block/