Add list of tags to Page objects
I'm only taking the slugs such as "tag-1", rather than WordPress's
"plain text" version of the tag (e.g. "tag 1"). I'm a bit surprised
WordPress has two representations of a tag, and that a tag in WordPress
can contain spaces.

In Astro, we'll only use the tag's "slug" in a page's front matter.