Move WordPress gallery knowledge to wp module
All the WordPress specific logic belongs in the wp module really, and
the format of the text that WordPress use to document the contents of a
gallery are very clearly "WordPress specific".

In order to move the logic out of the Page class, I've added support for
passing "filters" to Page that can be used to modify the HTML before it
gets converted to Markdown.