Pass HTML filters into Page.__init__()
The pressure to move the list of filters to the __init__ method is
coming from the fact that the Page.attachment_ids property currently
needs to know the format that WordPress is using to export the image
galleries. That knowledge really ought to be confined to the
GalleryFilter in the wp module.

If the filter is available as an instance variable on the Page then
we'll be able to read all of the image IDs (from inline <img> tags and
galleries) the same way; by looking at the `wp-image-` class attributes.