Remove awareness of WordPress galleries from Page
The page module doesn't need to know anything about the way WordPress
format the text that lists the images in a gallery. That code belongs in
the wp module.

By applying the GalleryFilter to the page content before retrieving the
list of attachment IDs, this commit can do away with the special
knowledge required to extract IDs from that syntax.