Insert line break after image at start of paragraph
WordPress's editor makes it easy for people to create paragraphs that start
with an image that then immediately run straight into some text, without
any white space. In general I don't feel that the image is intended to be
part of the paragraph that follows. This filter inserts a line break
between them.