Convert URLs for hosted images to relative paths
Images attached to pages on WordPress are identified in the export XML
as attachments, and loaded using an <img> tag that retrieves the image
from the WordPress domain.

In Astro those images will be stored in the same directory as the
Markdown file that references them, so can be loaded with a relative
path. The HostedImageFilter rewrites the URLs to relative filenames.

Note that it will log a warning if the markup contains an <img> tag that
refers to an attachment that isn't included in the XML export. A few
images were actually missing in the site that I'm working on as I write
this code.