Strip query string from image URLs
The URL of WordPress-hosted images are often adorned with query string
parameters like `?w=4096`, to set the size of the image that should be
retrieved.

In the modern era, it's far better to use srcset and to allow the client
to have some control over the most appropriate size to retrieve (based
on, amongst other things, the size of the viewport). Astro's `<Image/>`
tag provides us with some assistance in this department.

We can also benefit from those same smarts when loading local images out
of the `./src` directory, in Markdown files.

So, we don't need the query strings, and while they won't actually do
anything on a site served from Astro, let's keep things tidy and remove
them.