Remove links that navigate to wrapped image
Some of the pages in a site I'm moving off WordPress contain <a> tags
that wrap <img> tags.

That's no problem in itself, but the href attribute of these <a> tags
are set to the image that's displayed inside the link. At least for this
site, that's pointless.

I can't think of a great use case for linking directly to the image
that's contained within the link, so have added this filter to remove
the <a> tags.

Note that RemoveImageLinksFilter runs before IllustratedParagraphFilter,
which means that line breaks can be inserted between the image and the
text that tends to (on the site I'm working on) follow it immediately.