Add thumbnail image to Markdown front matter
WordPress can record which image was used to represent a post by setting
the _thumbnail_id metadata in its exported XML:

    <wp:postmeta>
      <wp:meta_key>_thumbnail_id</wp:meta_key>
      <wp:meta_value><![CDATA[1234]]></wp:meta_value>
    </wp:postmeta>

To parse that XML we're using XPath to find <wp:meta_value> elements
that follow a <wp:meta_key> element with the text of _thumbnail_id. The
XPath expression that we're using to do that depends on the ... syntax
for finding a node's parent.

See https://stackoverflow.com/a/33280875 for notes on that.

Also note that the `wp.posts()` function now takes a list of parsers.
I've introduced it so that parsers can be injected at run time. The main
reason for doing that now is that tests that don't concern themselves
with (for example) thumbnail images don't need to be updated to support
this new code. An alternative to adding this metadata XML to existing
tests' data would be to add behaviour to the production code so that
it can cope with missing data. But that's unrealistic, and makes the
code less clear. Injecting all the parsers at runtime, so we can choose
when the behaviour is active, is much nicer.

To put that another way, we get less coupling between code and tests,
and a simpler code base.