Use composition to inject tag parser
Now that parsing code can be injected into `wp.posts()` (a feature
introduced to help reduce the coupling in the parsing code, and between
test cases), I'm injecting the tag parsing function dynamically.

It causes less coupling between different parts of the code base. And
tests don't need to be setup with quite so much data in order to satisfy
parsers that don't concern them.